pyproject.toml
[tool.poetry]
name = "reader"
version = "0.1.0"
description = "A nice markdown reader"
authors = ["ddorn <diego.dorn@free.fr>"]
readme = "README.md"
license = "GPL-3.0"

[tool.poetry.scripts]
pucoti = "reader:app"

[tool.poetry.dependencies]
python = "^3.12"
pygame-ce = "^2.4.1"
typer = "^0.12.3"
marko = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
